Avoid allocation (and leak) in gmtime_r
This leak is rather hard to detect under lsan because if you call
`gmtime` before `gmtime_r` then the pointer is stashed in static data,
and if you call `gmtime_r` with stack allocation struct it gets found
there too (I think this is an lsan bug).

Regardlss, the leak is real and the fix is to simply avoid the
allocation.

Fixes: #15861